+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # v4.2
+## New features
 * Pathfinding using the A* algorithm is now possible!
 - Fixed omission which did not enable updating properties of a model when `model.properties` is a `struct`.
+- New function `multirun!` for running ensemble model simulations.
+
+## Deprecated
+- Keyword `replicates` of `run!` is deprecated in favor of `multirun!` and will be invalid in version 4.3.
+- `paramscan` with `replicates` is deprecated and will be invalid in version 4.3.
 
 # v4.1.2
 - Plotting with Plots.jl and `plotabm` is deprecated in favor of InteractiveDynamics.jl, Makie.jl and `abm_plot`.